create dag, rename claude folder
