+++ claude/settings.json
